+++ src/app/shared/models/aql/aql-ui.model.ts
@@ -15,60 +15,107 @@
  */
 
 import { LogicalOperator } from '../logical-operator.enum'
-import { IPhenotypeQueryApi } from '../phenotype/phenotype-query-api.interface'
 import { ConnectorNodeType } from 'src/app/shared/models/connector-node-type.enum'
 import { PARAMETER_REGEX } from '../../../core/constants/constants'
 import { ConnectorMainNodeUi } from '../connector-main-node-ui.interface'
-import { IAqlPhenotypeApi } from './aql-phenotype.interface'
+import { IAqlCohortApi } from './aql-cohort.interface'
 import { IUser } from '../user/user.interface'
+import { ICohortGroupApi } from '../project/cohort-group-api.interface'
+import { IAqlParameter } from './aql-parameter.interface'
+import { IDictionary } from '../dictionary.interface'
+import { AqlParameterOperator } from './aql-parameter-operator.type'
 
-export class AqlUiModel implements ConnectorMainNodeUi {
+export class AqlUiModel implements ConnectorMainNodeUi<ICohortGroupApi> {
+  private readonly OPERATOR_SUFFIX = '__OPERATOR'
+  private readonly NAME_SUFFIX = '__NAME'
+
   type: ConnectorNodeType.Aql
   id: number
   name: string
   query: string
+  queryWithOperatorPlaceholder: string
   isNegated: boolean
-  parameter: { name: string; value?: string }[]
+  parameters: IAqlParameter[]
   areParameterConfigured = true
   indexInGroup: number | null
   purpose: string
+  use: string
   owner?: IUser | null
 
-  constructor(aql: IAqlPhenotypeApi, isNegated: boolean = false) {
+  constructor(aql: IAqlCohortApi, isNegated: boolean = false) {
     this.type = ConnectorNodeType.Aql
     this.id = aql.id
     this.name = aql.name
     this.query = aql.query
     this.isNegated = isNegated
-    this.parameter = (aql.query.match(PARAMETER_REGEX) || []).map((name) => {
-      return { name, value: undefined }
-    })
     this.purpose = aql.purpose
+    this.use = aql.use
     this.owner = aql.owner
 
-    if (this.parameter.length) {
+    this.collectParameters()
+    if (this.parameters.length) {
       this.areParameterConfigured = false
     }
   }
 
-  public convertToApi(): IPhenotypeQueryApi {
+  private collectParameters(): void {
+    this.parameters = (this.query.match(PARAMETER_REGEX) || []).map((name) => {
+      return {
+        name,
+        value: undefined,
+        archetypeId: undefined,
+        operator: undefined,
+        path: undefined,
+      }
+    })
+
+    this.parameters.forEach((parameter) => {
+      const parameterPathRegex = new RegExp(`\\S+\\s+\\S+\\s+\\${parameter.name}`, 'gmi')
+      const fullParameterPath = this.query.match(parameterPathRegex)[0]
+      const fullParameterPathSplitted = fullParameterPath.split(' ')
+      const archetypeReferenceId = fullParameterPathSplitted[0].match(/\w+/)[0]
+      const archetypeIdRegex = new RegExp(`(?<=${archetypeReferenceId}\\[)(.+?)(?=\s*])`)
+
+      parameter.operator =
+        fullParameterPathSplitted[1] in AqlParameterOperator
+          ? (fullParameterPathSplitted[1] as AqlParameterOperator)
+          : AqlParameterOperator['!=']
+      parameter.path = fullParameterPathSplitted[0].split(archetypeReferenceId)[1]
+      parameter.archetypeId = this.query.match(archetypeIdRegex)[0]
+
+      const pathWithInjectedPlaceholder = fullParameterPath
+        .replace(parameter.name, parameter.name + this.NAME_SUFFIX)
+        .replace(parameter.operator, parameter.name + this.OPERATOR_SUFFIX)
+
+      this.queryWithOperatorPlaceholder = (this.queryWithOperatorPlaceholder
+        ? this.queryWithOperatorPlaceholder
+        : this.query
+      ).replace(fullParameterPath, pathWithInjectedPlaceholder)
+    })
+  }
+
+  public convertToApi(): ICohortGroupApi {
     return this.isNegated ? this.convertToNegatedApiGroup() : this.getAqlForApi()
   }
 
-  private getAqlForApi(): IPhenotypeQueryApi {
+  private getAqlForApi(): ICohortGroupApi {
     return {
       type: ConnectorNodeType.Aql,
-      aql: {
+      parameters: this.convertParametersToApi(),
+      query: {
         id: this.id,
         name: this.name,
-        query: this.insertParamsForApi(this.query),
+        query: this.parameters.length
+          ? this.insertOperatorForApi(this.queryWithOperatorPlaceholder)
+          : this.query,
         purpose: this.purpose,
+        use: this.use,
         owner: this.owner,
       },
     }
   }
 
-  private convertToNegatedApiGroup(): IPhenotypeQueryApi {
+  private convertToNegatedApiGroup(): ICohortGroupApi {
     return {
       type: ConnectorNodeType.Group,
       operator: LogicalOperator.Not,
@@ -76,10 +123,19 @@ export class AqlUiModel implements ConnectorMainNodeUi {
     }
   }
 
-  private insertParamsForApi(queryString: string): string {
+  private convertParametersToApi(): IDictionary<string, string> {
+    return this.parameters.reduce((dictionary, parameter) => {
+      dictionary[parameter.name] = parameter.value
+      return dictionary
+    }, {})
+  }
+
+  private insertOperatorForApi(queryString: string): string {
     let resultString = queryString
-    this.parameter.forEach((param) => {
-      resultString = resultString.replace(param.name, param.value ? param.value : param.name)
+    this.parameters.forEach((parameter) => {
+      resultString = resultString
+        .replace(parameter.name + this.NAME_SUFFIX, parameter.name)
+        .replace(parameter.name + this.OPERATOR_SUFFIX, parameter.operator)
     })
     return resultString
   }