+++ src/app/shared/models/connector-node-type.enum.ts
@@ -16,7 +16,6 @@
 
 export enum ConnectorNodeType {
   Group = 'GROUP',
-  Phenotype = 'PHENOTYPE',
   Aql = 'AQL',
   Aqb_Composition = 'AQB_COMPOSITION',
   Aqb_Item = 'AQB_ITEM',